separate some functions to individual files
