+++ .gitignore
@@ -8,7 +8,7 @@
 *.vscode
 build
 cmake
-CMake*
+CMakePresets.json
 Drivers
 
 # Prerequisites
