+++ Core/Inc/led.h
@@ -1,0 +1,4 @@
+
+void led_on(const int num);
+void led_off(const int num);
+void led_set(const int bits, const int mask);
